backend/pyproject.toml
[project]
name = "resume-to-latex-backend"
version = "0.1.0"
description = "Resume to LaTeX converter backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.2",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
]
